DDFLSBP-273 - Moved dashboard and reservation list logic
