+++ web/modules/custom/dpl_dashboard/dpl_dashboard.module
@@ -4,3 +4,51 @@
  * @file
  * DPL dashboard drupal module file.
  */
+
+use Drupal\Core\Form\FormStateInterface;
+use Drupal\dpl_dashboard\DplDashboardSettings;
+
+/**
+ * Implements hook_form_alter().
+ */
+function dpl_dashboard_form_dpl_library_agency_list_size_settings_form_alter (array &$form, FormStateInterface $form_state): void
+{
+  $config = \Drupal::service('config.factory')
+    ->getEditable('dpl_library_agency.list_size_settings');
+
+  $form['dashboard_list_size_settings'] = [
+    '#type' => 'fieldset',
+    '#title' => t('Dashboard list size settings', [], ['context' => 'List size (settings)']),
+    '#description' => t('The number of items to display in the dashboard list.', [], ['context' => 'List size (settings)']),
+    '#tree' => FALSE,
+  ];
+
+  $form['dashboard_list_size_settings']['dashboard_list_size_desktop'] = [
+    '#type' => 'number',
+    '#title' => t('Dashboard list size on desktop', [], ['context' => 'List size (settings)']),
+    '#default_value' => $config->get('dashboard_list_size_desktop') ?? DplDashboardSettings::DASHBOARD_LIST_SIZE_DESKTOP,
+    '#min' => 1,
+    '#step' => 1,
+  ];
+
+  $form['dashboard_list_size_settings']['dashboard_list_size_mobile'] = [
+    '#type' => 'number',
+    '#title' => t('Dashboard list size on mobile', [], ['context' => 'List size (settings)']),
+    '#default_value' => $config->get('dashboard_list_size_mobile') ?? DplDashboardSettings::DASHBOARD_LIST_SIZE_MOBILE,
+    '#min' => 1,
+    '#step' => 1,
+  ];
+
+  $form['#submit'][] = 'dpl_dashboard_handler_submit';
+}
+
+/**
+ * Custom submit callback.
+ */
+function dpl_dashboard_handler_submit(array &$form, FormStateInterface $form_state): void {
+  \Drupal::service('config.factory')
+    ->getEditable('dpl_library_agency.list_size_settings')
+    ->set('dashboard_list_size_desktop', $form_state->getValue('dashboard_list_size_desktop'))
+    ->set('dashboard_list_size_mobile', $form_state->getValue('dashboard_list_size_mobile'))
+    ->save();
+}
