+++ web/modules/custom/dpl_dashboard/dpl_dashboard.routing.yml
@@ -8,18 +8,7 @@ dpl_dashboard.list:
 dpl_dashboard.settings:
   path: "/admin/config/dpl-library-agency/dashboard-settings"
   defaults:
-    _form: '\Drupal\dpl_dashboard\Form\DashboardSettingsForm'
     _title: "Dashboard settings"
-  requirements:
-    _permission: "administer site configuration"
-  options:
-    _admin_route: TRUE
-
-dpl_list_size.settings:
-  path: "/admin/config/dpl-library-agency/list-size-settings"
-  defaults:
-    _title: "List size settings"
-    _form: '\Drupal\dpl_dashboard\Form\DashboardListSizeSettingsForm'
   requirements:
     _permission: "administer site configuration"
   options:
