+++ web/modules/custom/dpl_dashboard/src/DplDashboardSettings.php
@@ -9,8 +9,8 @@ use Drupal\dpl_react\DplReactConfigBase;
  */
 class DplDashboardSettings extends DplReactConfigBase {
 
-  const PAGE_SIZE_DESKTOP = 25;
-  const PAGE_SIZE_MOBILE = 25;
+  const DASHBOARD_LIST_SIZE_DESKTOP = 25;
+  const DASHBOARD_LIST_SIZE_MOBILE = 15;
 
   /**
    * Gets the configuration key for dashboard settings.
