--- web/modules/custom/dpl_dashboard/src/Form/DashboardSettingsForm.php
@@ -1,100 +1,0 @@
-<?php
-
-namespace Drupal\dpl_dashboard\Form;
-
-use Drupal\Core\Config\ConfigFactoryInterface;
-use Drupal\Core\Form\ConfigFormBase;
-use Drupal\Core\Form\FormStateInterface;
-use Drupal\dpl_dashboard\DplDashboardSettings;
-use Drupal\dpl_react\DplReactConfigInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-
-/**
- * Dashboard list setting form.
- */
-class DashboardSettingsForm extends ConfigFormBase {
-
-  /**
-   * Default constructor.
-   *
-   * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
-   *   Config factory.
-   * @param \Drupal\dpl_react\DplReactConfigInterface $configService
-   *   Reservation list configuration object.
-   */
-  public function __construct(
-    ConfigFactoryInterface $config_factory,
-    protected DplReactConfigInterface $configService,
-  ) {
-    parent::__construct($config_factory);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container): self {
-    return new static(
-      $container->get('config.factory'),
-      \Drupal::service('dpl_dashboard.settings')
-    );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function getEditableConfigNames(): array {
-    return [
-      $this->configService->getConfigKey(),
-    ];
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getFormId(): string {
-    return 'dashboard_settings_form';
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function buildForm(array $form, FormStateInterface $form_state): array {
-    $config = $this->configService->loadConfig();
-
-    $form['settings'] = [
-      '#type' => 'fieldset',
-      '#title' => $this->t('Basic settings', [], ['context' => 'Dashboard (settings)']),
-      '#tree' => FALSE,
-    ];
-
-    $form['settings']['page_size_mobile'] = [
-      '#type' => 'number',
-      '#title' => $this->t('Page size mobile', [], ['context' => 'Dashboard (settings)']),
-      '#default_value' => $config->get('page_size_mobile') ?? DplDashboardSettings::PAGE_SIZE_MOBILE,
-      '#min' => 0,
-      '#step' => 1,
-    ];
-
-    $form['settings']['page_size_desktop'] = [
-      '#type' => 'number',
-      '#title' => $this->t('Page size desktop', [], ['context' => 'Dashboard (settings)']),
-      '#default_value' => $config->get('page_size_desktop') ?? DplDashboardSettings::PAGE_SIZE_DESKTOP,
-      '#min' => 0,
-      '#step' => 1,
-    ];
-
-    return parent::buildForm($form, $form_state);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function submitForm(array &$form, FormStateInterface $form_state): void {
-    parent::submitForm($form, $form_state);
-    $this->config($this->configService->getConfigKey())
-      ->set('page_size_desktop', $form_state->getValue('page_size_desktop'))
-      ->set('page_size_mobile', $form_state->getValue('page_size_mobile'))
-      ->save();
-  }
-
-}
