+++ web/modules/custom/dpl_library_agency/dpl_library_agency.links.menu.yml
@@ -3,8 +3,15 @@ dpl_library_agency.settings:
   route_name: dpl_library_agency.settings
   parent: system.admin_config
   weight: 0
+
 dpl_library_agency.general_settings_form:
   title: "General Settings"
   description: "Administer various settings for a library agency."
   route_name: dpl_library_agency.general_settings
+  parent: dpl_library_agency.settings
+
+dpl_library_agency.list_size_settings_form:
+  title: "List Size Settings"
+  description: "Administer list size settings for a library agency."
+  route_name: dpl_library_agency.list_size_settings
   parent: dpl_library_agency.settings
