--- web/modules/custom/dpl_list_size/dpl_list_size.info.yml
@@ -1,5 +1,0 @@
-name: 'DPL List Size'
-type: module
-description: 'Module for handling various list size configuration across the site.'
-package: DPL
-core_version_requirement: ^10
