--- web/modules/custom/dpl_list_size/dpl_list_size.module
@@ -1,6 +1,0 @@
-<?php
-
-/**
- * @file
- * Primary module hooks for DPL List Size module.
- */
