--- web/modules/custom/dpl_list_size/dpl_list_size.routing.yml
@@ -1,10 +1,0 @@
-dpl_list_size.settings:
-  path: "/admin/config/dpl-library-agency/list-size-settings"
-  defaults:
-    _title: "List size settings"
-    _form: '\Drupal\dpl_list_size\Form\ListSizeSettingsForm'
-  requirements:
-    _permission: "administer site configuration"
-  options:
-    _admin_route: TRUE
-
