--- web/modules/custom/dpl_list_size/dpl_list_size.services.yml
@@ -1,4 +1,0 @@
-services:
-  dpl_list_size.settings:
-    class: Drupal\dpl_list_size\DplListSizeSettings
-    arguments: ["@config.manager"]
