+++ web/modules/custom/dpl_reservations/dpl_reservations.links.menu.yml
@@ -3,9 +3,3 @@ dpl_reservations.list:
   menu_name: user-profile-menu
   route_name: dpl_reservations.list
   description: "List user reservations"
-
-dpl_reservations.settings_form:
-  title: "Reservation list settings"
-  route_name: dpl_reservations.settings
-  description: "Change reservation list settings"
-  parent: dpl_library_agency.settings
