+++ web/modules/custom/dpl_reservations/dpl_reservations.module
@@ -8,6 +8,8 @@
  */
 
 use Drupal\Core\Url;
+use Drupal\Core\Form\FormStateInterface;
+use Drupal\dpl_reservations\DplReservationsSettings;
 
 /**
  * Implements hook_dpl_react_apps_data().
@@ -21,4 +23,51 @@ function dpl_reservations_dpl_react_apps_data(array &$data): void {
   $data['texts'] += [
     'reservation-list-loan-before' => t('Borrow before @date', [], ['context' => 'Reservation list']),
   ];
+}
+
+/**
+ * Implements hook_form_alter().
+ */
+function dpl_reservations_form_dpl_library_agency_list_size_settings_form_alter (array &$form, FormStateInterface $form_state): void
+{
+  $config = \Drupal::service('config.factory')
+    ->getEditable('dpl_library_agency.list_size_settings');
+
+  var_dump($config->get('reservation_list_size_desktop'));
+
+  $form['reservation_list_size_settings'] = [
+    '#type' => 'fieldset',
+    '#title' => t('Reservation list size settings', [], ['context' => 'List size (settings)']),
+    '#description' => t('The number of items to display in the reservation list.', [], ['context' => 'List size (settings)']),
+    '#tree' => FALSE,
+  ];
+
+  $form['reservation_list_size_settings']['reservation_list_size_desktop'] = [
+    '#type' => 'number',
+    '#title' => t('Reservation list size on desktop', [], ['context' => 'List size (settings)']),
+    '#default_value' => $config->get('reservation_list_size_desktop') ?? DplReservationsSettings::RESERVATION_LIST_SIZE_DESKTOP,
+    '#min' => 1,
+    '#step' => 1,
+  ];
+
+  $form['reservation_list_size_settings']['reservation_list_size_mobile'] = [
+    '#type' => 'number',
+    '#title' => t('Reservation list size on mobile', [], ['context' => 'List size (settings)']),
+    '#default_value' => $config->get('reservation_list_size_mobile') ?? DplReservationsSettings::RESERVATION_LIST_SIZE_MOBILE,
+    '#min' => 1,
+    '#step' => 1,
+  ];
+
+  $form['#submit'][] = 'dpl_reservations_handler_submit';
+}
+
+/**
+ * Custom submit callback.
+ */
+function dpl_reservations_handler_submit(array &$form, FormStateInterface $form_state): void {
+  \Drupal::service('config.factory')
+    ->getEditable('dpl_library_agency.list_size_settings')
+    ->set('reservation_list_size_desktop', $form_state->getValue('reservation_list_size_desktop'))
+    ->set('reservation_list_size_mobile', $form_state->getValue('reservation_list_size_mobile'))
+    ->save();
 }
