+++ web/modules/custom/dpl_reservations/src/DplReservationsSettings.php
@@ -8,8 +8,8 @@ use Drupal\dpl_react\DplReactConfigBase;
  * Class that handles reservations settings.
  */
 class DplReservationsSettings extends DplReactConfigBase {
-  const PAGE_SIZE_DESKTOP = 25;
-  const PAGE_SIZE_MOBILE = 25;
+  const RESERVATION_LIST_SIZE_DESKTOP = 25;
+  const RESERVATION_LIST_SIZE_MOBILE = 25;
 
   /**
    * Gets the configuration key for reservation settings.
